+++ pyxelate/__init__.py
@@ -1,8 +1,6 @@
-__version__ = "2.1.2"
+__version__ = "2.1.3"
 
-__short_description__ = (
-    "Downsample images to 8-bit pixel art.",
-)
+__short_description__ = ("Downsample images to 8-bit pixel art.",)
 __license__ = "MIT"
 __author__ = "Richard Nagyfi"
 __github_username__ = "sedthh"